refactor(helpers): refactor Loggable class and add base_classes module
The Loggable class in the helpers module has been refactored to improve its functionality and readability. The class now inherits from a new LoggableDescriptor class defined in the base_classes module. This descriptor allows for easy access to a logger specific to a class method. Additionally, a new _get_parent_logging_device() function has been added to determine the parent logging device by inspecting the caller's log_device or parent_log_device attribute.

These changes improve the logging capabilities of the Loggable class and enhance consistency in logging behavior and hierarchy.
