+++ inspy_logger/__init__.py
@@ -1,9 +1,11 @@
 #!/usr/bin/env python3
 import inspect
+import os
 import logging
 from rich.logging import RichHandler
 from inspy_logger.__about__ import __prog__ as PROG_NAME
 from inspy_logger.helpers import find_variable_in_call_stack
+
 
 # Let's set up some constants.
 LEVEL_MAP = [
@@ -144,8 +146,7 @@ class Logger:
                 child.set_level(file_level=file_level)
 
     def get_child(self, name=None, console_level=None, file_level=None):
-        self.logger.debug("Getting child logger")
-        console_level = console_level or DEFAULT_LOGGING_LEVEL
+        console_level = console_level or self.__console_level
         caller_frame = inspect.stack()[1]
 
         if name is None:
@@ -159,7 +160,7 @@ class Logger:
             if child.logger.name == child_logger_name:
                 return child
 
-        child_logger = Logger(child_logger_name, console_level, file_level, parent=self)
+        child_logger = Logger(name=child_logger_name, console_level=console_level, file_level=file_level, parent=self)
         self.children.append(child_logger)
         return child_logger
 
@@ -195,6 +196,19 @@ class Logger:
 
         return results
 
+    def debug(self, message):
+        self.logger.debug(message)
+
+    def info(self, message):
+        self.logger.info(message)
+
+    def warning(self, message):
+        self.logger.warning(message)
+
+    def error(self, message):
+        self.logger.error(message)
+
+
     def __repr__(self):
         name = self.name
         hex_id = hex(id(self))
@@ -208,7 +222,47 @@ class Logger:
         if parent_part.endswith('|'):
             parent_part = str(parent_part[:-2])
 
-        return f'<Logger: {name} at {hex_id}{parent_part}>'
+        return f'<Logger: {name} w/ level {self.logger.level} at {hex_id}{parent_part}>'
+
+
+    @classmethod
+    def create_logger_for_caller(cls):
+        """
+        Creates a logger for the module that calls this method.
+
+        Returns:
+            Logger: An instance of the Logger class for the calling module.
+        """
+        frame = inspect.currentframe().f_back
+        #print(frame)
+        #print(dir(frame))
+        #print(dir(inspect.getmodule(frame)))
+        #print(inspect.getmodule(frame).__name__)
+        module_path = cls._determine_module_path(frame)
+        if module_path:
+            return cls(module_path)
+        else:
+            raise ValueError("Unable to determine module path for logger creation.")
+
+    @staticmethod
+    def _determine_module_path(frame):
+        """
+        Determines the in-project path of the module from the call frame.
+
+        Args:
+            frame: The frame from which to determine the module path.
+
+        Returns:
+            str: The in-project path of the module.
+        """
+        module = inspect.getmodule(frame)
+        if module:
+            base_path = os.path.dirname(os.path.abspath(module.__file__))
+            relative_path = os.path.relpath(frame.f_code.co_filename, base_path)
+            module_path = relative_path.replace(os.path.sep, '.').rstrip('.py')
+            return module_path
+        return None
+
 
 
 found_level = find_variable_in_call_stack('INSPY_LOG_LEVEL', DEFAULT_LOGGING_LEVEL)
@@ -223,104 +277,4 @@ add_child = LOG_DEVICE.get_child
 
 InspyLogger = Logger
 
-def _get_parent_logging_device():
-    """
-    Determines the parent logging device by inspecting the caller's log_device or parent_log_device attribute.
-
-    Returns:
-        Logger: The parent logging device.
-    """
-    MOD_LOGGER.debug("Determining parent logging device")
-    caller_frame = inspect.currentframe().f_back
-    caller_locals = caller_frame.f_locals
-
-    if "logger" in caller_locals:
-        return caller_locals["logger"]
-    elif "parent_log_device" in caller_locals:
-        return caller_locals["parent_log_device"]
-    else:
-        raise ValueError("Unable to determine the parent logging device.")
-
-
-class Loggable:
-    """
-    A metaclass to enhance classes with logging capabilities. Classes that inherit from
-    'Loggable' can instantly access a logger without manually setting it up. This logger
-    is derived from a parent logger, ensuring consistent logging behavior and hierarchy.
-
-    Attributes:
-        - log_device: The logger device associated with the instance of the class.
-    """
-
-    def __init__(self, parent_log_device=None, **kwargs):
-        self.parent_log_device = parent_log_device
-        self.__log_name = self.__class__.__name__
-        if self.parent_log_device is not None:
-            self.__log_device = self.parent_log_device.get_child(
-                self.__class__.__name__
-            )
-        else:
-            self.__log_device = _get_parent_logging_device().get_child(
-                self.__class__.__name__
-            )
-
-    @property
-    def log_device(self):
-        return self.__log_device
-
-    @log_device.setter
-    def log_device(self, new):
-        if not isinstance(new, Logger):
-            raise TypeError('log_device must be of type "Logger"')
-
-        self.__log_device = new
-
-    def create_child_logger(self, name=None, override=False):
-        """
-        Creates and returns a child logger of this object's logger.
-
-        Parameters:
-            name (str, optional): The name of the child logger.
-                If not provided, the name of the calling function is used.
-            override (bool, optional): A flag to override the membership check. Defaults to False.
-
-        Returns:
-            Logger: An instance of the Logger class that represents the child logger.
-        """
-        if not override:
-            self.__is_member__()
-
-        if name is None:
-            name = inspect.stack()[1][
-                3
-            ]  # Get the name of the calling function if no name is provided
-
-        return self.log_device.get_child(name)
-
-    def __is_member__(self):
-        """
-        Checks whether the caller of this method is a member of the same class.
-
-        Raises:
-            PermissionError: If the caller of this method is not a member of the same class.
-        """
-        log_device = self.log_device.get_child("__is_member__")
-        log = log_device.logger
-
-        current_frame = inspect.currentframe()
-        log.debug(f"Current frame: {current_frame}")
-
-        caller_frame = current_frame.f_back
-        log.debug(f"Caller frame: {caller_frame}")
-
-        caller_self = caller_frame.f_locals.get("self", None)
-        log.debug(f"Caller self: {caller_self}")
-
-        log.debug("Checking if caller is a member of this class...")
-        if not isinstance(caller_self, self.__class__):
-            raise PermissionError(
-                "Access denied.\n"
-                f"Method can only be accessed by members of the same class. {caller_self.__class__.__name__} is not such a member"
-            )
-
-        log.debug(f"Access granted to {caller_self.__class__.__name__}")
+from inspy_logger.helpers.base_classes import Loggable
